octavo commit

+++ blueprints/ingresos/__init__.py
@@ -1,7 +1,12 @@
-from flask import Blueprint, render_template, request, redirect, url_for
+from flask import Blueprint, render_template, request, redirect, url_for,session
 import sqlite3
 
 ingresos_bp = Blueprint('ingresos', __name__, template_folder='templates')
+
+def verifica():
+    if 'cargo' not in session or session['cargo'] != "Administrador":
+        return False
+    return True
 
 @ingresos_bp.route("/ingresos")
 def index_ingreso():
@@ -15,6 +20,8 @@ def index_ingreso():
 
 @ingresos_bp.route("/crear_i", methods=["GET", "POST"])
 def crear_i():
+    if not verifica():
+        return redirect(url_for('ingresos.index_ingreso'))
     if request.method == "POST":
         # Recupera los datos del formulario
         ci_afiliado = request.form["ci_afiliado"]
@@ -29,7 +36,7 @@ def crear_i():
         try:
             cursor.execute("""
                 INSERT INTO ingreso (ci_afiliado, monto, estado, fecha, observaciones) 
-                VALUES (?, ?, ?, ?, ?, ?, ?)
+                VALUES (?, ?, ?, ?, ?)
             """, (ci_afiliado, monto, estado, fecha, observaciones))
             conn.commit()
         except sqlite3.IntegrityError as e:
@@ -38,7 +45,7 @@ def crear_i():
         finally:
             conn.close()
         
-        return redirect(url_for('ingresos.index'))
+        return redirect(url_for('ingresos.index_ingreso'))
     
     # Obtener afiliados para el formulario
     conn = sqlite3.connect("star_service.db")
@@ -53,6 +60,8 @@ def crear_i():
 
 @ingresos_bp.route("/editar_i/<int:id>", methods=["GET", "POST"])
 def editar_i(id):
+    if not verifica():
+        return redirect(url_for('ingresos.index_ingreso'))
     conn = sqlite3.connect("star_service.db")
     conn.row_factory = sqlite3.Row
     cursor = conn.cursor()
@@ -82,22 +91,24 @@ def editar_i(id):
         finally:
             conn.close()
 
-        return redirect(url_for('ingresos.index'))
+        return redirect(url_for('ingresos.index_ingreso'))
 
     # Obtener la lista de afiliados para el formulario
     cursor.execute("SELECT ci, nombres FROM afiliado")
     afiliados = cursor.fetchall()
     conn.close()
 
-    return render_template("editar.html", ingreso=ingreso, afiliados=afiliados)
+    return render_template("editar_i.html", ingreso=ingreso, afiliados=afiliados)
 
 
 @ingresos_bp.route("/eliminar_i/<int:id>")
 def eliminar_i(id):
+    if not verifica():
+        return redirect(url_for('ingresos.index_ingreso'))
     conn = sqlite3.connect("star_service.db")
     cursor = conn.cursor()
     cursor.execute("DELETE FROM ingreso WHERE id = ?", (id,))
     conn.commit()
     conn.close()
-    return redirect(url_for('ingresos.index'))
+    return redirect(url_for('ingresos.index_ingreso'))
